Two small but important fixes.
First, the list of line numbers is memoized into an array
to quickly determine if the line number has been duplicated.

Second, the ord_header.ord_id value is only set on the ord_import
record if we know for sure the order was successfully persisted
to the database. If this was not in place, the update to ord_import
would fail causing the entire Importer to bomb out.